corrigindo review

+++ gamelog/apps/review/models.py
@@ -6,7 +6,7 @@ class Review(models.Model):
     name = models.CharField('Nome', max_length=50)
     description = models.TextField('Descricao', max_length=100)
     review = models.CharField('Review', max_length=500) 
-    rating = models.IntegerField('Nota')
+    rating = models.FloatField('Nota')
     release_date = models.DateField('Data de Lançamento', auto_now=False, auto_now_add=False)
 
     class Meta:
